[Blue] Chapter 5. JPA 활용

+++ SSU/UMC/Spring_Boot_B/src/main/java/com/ssu/umc9th2/spring_boot_b/domain/review/entity/Review.java
@@ -1,11 +1,10 @@
 package com.ssu.umc9th2.spring_boot_b.domain.review.entity;
 
-import com.ssu.umc9th2.common.base.BaseEntity;
+import com.ssu.umc9th2.spring_boot_b.common.base.BaseEntity;
 import com.ssu.umc9th2.spring_boot_b.domain.restaurant.entity.Restaurant;
 import com.ssu.umc9th2.spring_boot_b.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
-import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -26,7 +25,7 @@ public class Review extends BaseEntity {
     @JoinColumn(name = "restaurant_id")
     private Restaurant restaurant;
 
-    private BigDecimal rating;
+    private Double rating;
 
     private String content;
 }
